Update everything to typescript
