+++ package.json
@@ -4,8 +4,14 @@
   "description": "A view component for react-native with pinch to zoom, tap to move and double tap to zoom capability.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "format": "prettier --write \"src/**/*.ts\" \"src/**/*.js\"",
+    "version": "npm run format && git add -A src",
+    "postversion": "git push && git push --tags"
   },
+  "prepare": "npm run build",
+  "files": ["lib/**/*"],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/DuDigital/react-native-zoomable-view.git"
@@ -34,6 +40,14 @@
   "homepage": "https://github.com/DuDigital/react-native-zoomable-view#readme",
   "types": "typings/index.d.ts",
   "dependencies": {
-    "prop-types": "^15.5.10"
+    "prop-types": "^15.7.2",
+    "react-native": ">=0.54.0"
+  },
+  "devDependencies": {
+    "@types/react": "^17.0.11",
+    "@types/react-dom": "^17.0.7",
+    "prettier": "^2.3.1",
+    "tslint": "^6.1.3",
+    "tslint-config-prettier": "^1.18.0"
   }
 }
