+++ src/typings/index.d.ts
@@ -26,8 +26,11 @@ declare module '@dudigital/react-native-zoomable-view' {
   }
 
   export interface ReactNativeZoomableViewProps extends ViewProps {
+    // options
     zoomEnabled?: boolean;
     initialZoom?: number;
+    initialOffsetX?: number;
+    initialOffsetY?: number;
     maxZoom?: number;
     minZoom?: number;
     doubleTapDelay?: number;
@@ -38,27 +41,46 @@ declare module '@dudigital/react-native-zoomable-view' {
     pinchToZoomOutSensitivity?: number;
     zoomCenteringLevelDistance?: number;
     movementSensibility?: number;
-    initialOffsetX?: number;
-    initialOffsetY?: number;
     longPressDuration?: number;
     captureEvent?: boolean;
+    style?: any;
+
+    // callbacks
     onDoubleTapBefore?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
     onDoubleTapAfter?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
     onShiftingBefore?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => boolean;
-    onShiftingAfter?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
+    onShiftingAfter?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => boolean;
     onShiftingEnd?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
     onZoomBefore?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => boolean | void;
     onZoomAfter?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
     onZoomEnd?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
-    onLongPress?: (event: Event, gestureState: PanResponderGestureState) => void;
+    onLongPress?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
     onStartShouldSetPanResponder?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent, baseComponentResult: boolean) => boolean;
     onMoveShouldSetPanResponder?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent, baseComponentResult: boolean) => boolean;
     onPanResponderGrant?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
     onPanResponderEnd?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
-    onPanResponderMove?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => void;
+    onPanResponderMove?: (event: Event, gestureState: PanResponderGestureState, zoomableViewEventObject: ZoomableViewEvent) => boolean;
   }
 
-  export class ReactNativeZoomableView extends React.PureComponent<ReactNativeZoomableViewProps> {}
+  export interface ReactNativeZoomableViewState {
+    offsetX: number;
+    offsetY: number;
+    lastX: number;
+    lastY: number;
+    lastZoomLevel: number;
+    zoomLevel: number;
+    originalWidth: number;
+    originalHeight: number;
+    lastMovePinch: boolean;
+  }
+
+  export class ReactNativeZoomableView extends React.PureComponent<ReactNativeZoomableViewProps> {
+    // events
+    zoomTo: (newZoomLevel: number, bindToBorders: boolean) => Promise<void>;
+    zoomBy: (zoomLevelChange: number, bindToBorders: boolean) => Promise<void>;
+    moveTo: (newOffsetX: number, newOffsetY: number, bindToBorders: boolean) => Promise<boolean>;
+    moveBy: (offsetChangeX: number, offsetChangeY: number, bindToBorders: boolean) => Promise<boolean>;
+  }
 
   export interface ReactNativeZoomableViewWithGesturesProps extends ReactNativeZoomableViewProps {
     swipeLengthThreshold?: number,
